+++ test/IntrinsicPlasticityTestCase.cpp
@@ -28,26 +28,26 @@ void IntrinsicPlasticityTestCase::learn()
 
   ip.initialize();
   Eigen::VectorXd p = ip.currentParameters();
-  ASSERT_EQUALS_DELTA((double) p(0), (double) 1.0, (double) 1e-3);
-  ASSERT_EQUALS_DELTA((double) p(1), (double) 1.0, (double) 1e-3);
-  ASSERT_NOT_EQUALS((double) p(2), (double) 0.0);
-  ASSERT_NOT_EQUALS((double) p(3), (double) 0.0);
+  ASSERT_EQUALS_DELTA(p(0), 1.0, 1e-3);
+  ASSERT_EQUALS_DELTA(p(1), 1.0, 1e-3);
+  ASSERT_NOT_EQUALS(p(2), 0.0);
+  ASSERT_NOT_EQUALS(p(3), 0.0);
   p(2) = 1e-3;
   p(3) = 1e-3;
   ip.setParameters(p);
   p = ip.currentParameters();
-  ASSERT_EQUALS_DELTA((double) p(0), (double) 1.0, (double) 1e-3);
-  ASSERT_EQUALS_DELTA((double) p(1), (double) 1.0, (double) 1e-3);
-  ASSERT_NOT_EQUALS((double) p(2), (double) 0.0);
-  ASSERT_NOT_EQUALS((double) p(3), (double) 0.0);
+  ASSERT_EQUALS_DELTA(p(0), 1.0, 1e-3);
+  ASSERT_EQUALS_DELTA(p(1), 1.0, 1e-3);
+  ASSERT_NOT_EQUALS(p(2), 0.0);
+  ASSERT_NOT_EQUALS(p(3), 0.0);
 
   Eigen::VectorXd y(2);
   y.fill(0.0);
   for(int i = 0; i < samples; i++)
     y += ip(X.col(i));
   Eigen::VectorXd mean = y / (double) samples;
-  ASSERT_EQUALS_DELTA((double) mean(0), (double) 0.5, (double) 1e-2);
-  ASSERT_EQUALS_DELTA((double) mean(1), (double) 0.5, (double) 1e-2);
+  ASSERT_EQUALS_DELTA(mean(0), 0.5, 1e-2);
+  ASSERT_EQUALS_DELTA(mean(1), 0.5, 1e-2);
   const double e = ip.error();
 
   OpenANN::MBSGD sgd(5e-5, 0.9, 1);
@@ -61,7 +61,7 @@ void IntrinsicPlasticityTestCase::learn()
   for(int i = 0; i < samples; i++)
     y += ip(X.col(i));
   mean = y / (double) samples;
-  ASSERT_EQUALS_DELTA((double) mean(0), (double) 0.2, (double) 1e-2);
-  ASSERT_EQUALS_DELTA((double) mean(1), (double) 0.2, (double) 1e-2);
+  ASSERT_EQUALS_DELTA(mean(0), 0.2, 1e-2);
+  ASSERT_EQUALS_DELTA(mean(1), 0.2, 1e-2);
   ASSERT(e > ip.error());
 }
